pyproject.toml
[project]
name = "root"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework>=3.15.2",
    "pygments>=2.19.1",
    "wagtail>=6.3.2",
]

[dependency-groups]
dev = [
    "mypy>=1.14.0",
    "pre-commit>=4.0.1",
    "ruff>=0.8.4",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "A", # flake8-builtins
    "ANN", # flake8-annotations
    "B", # flake8-bugbear
    "C", # Pylint-Convention
    "COM", # flake8-commas
    "E", # pycodestyle - Error
    "F", # Pyflakes
    "I", # isort
    "N", # pep8-naming
    "S", # flake8-bandit
    "UP", # pyupgrade
    "W", # pycodestyle - Warning
]
ignore = [
    "ANN002", # Missing type annotation for *args
    "ANN003", # Missing type annotation for **kwargs
    "ANN101", # Missing type annotation for self in method
    "COM812", # missing-trailing-comma
    "D105", # Missing docstring in magic method
    "D107", # Missing docstring in __init__
    "D203", # 1 blank line required before class docstring (conflicts with D211)
    "D213", # Multi-line docstring symmary should start at the second line (conflicts with D212)
    "D301", # Use r if any backslashes in a docstring
    "E203", # Whitespace before ':'
]

[tool.ruff.lint.per-file-ignores]
"**/tests.py" = ["S101"]
"**/migrations/*" = ["ANN201", "ANN001", "E501", "N806"]

[tool.mypy]
exclude = [
    "/migrations/"
]

[[tool.mypy.overrides]]
module = [
    "django.apps",
    "django.conf.urls.static",
    "django.conf",
    "django.contrib.auth.models",
    "django.contrib.staticfiles.urls",
    "django.contrib",
    "django.core.handlers.wsgi",
    "django.core.management",
    "django.core.paginator",
    "django.core.wsgi",
    "django.db.models.deletion",
    "django.db.models",
    "django.db",
    "django.template.response",
    "django.urls",
    "django",
    "modelcluster.contrib.taggit",
    "modelcluster.contrib",
    "modelcluster.fields",
    "modelcluster",
    "pygments",
    "pygments.formatters.html",
    "pygments.lexers",
    "pygments.styles",
    "rest_framework.decorators",
    "rest_framework.response",
    "rest_framework.routers",
    "rest_framework.serializers",
    "rest_framework",
    "taggit.models",
    "wagtail.admin.panels",
    "wagtail.admin",
    "wagtail.blocks",
    "wagtail.contrib.forms.models",
    "wagtail.contrib.forms.panels",
    "wagtail.contrib.settings.models",
    "wagtail.documents",
    "wagtail.embeds.blocks",
    "wagtail.fields",
    "wagtail.images.blocks",
    "wagtail.models",
    "wagtail.search",
    "wagtail.snippets.models",
    "wagtail.snippets.views.snippets",
    "wagtail",
]
ignore_missing_imports = true
